+++ src/adt/mult/iterativemult.jl
@@ -30,12 +30,12 @@ end
 
 function iterativemult(x::ADT, y::ADT, alg::DMRGAlgorithm)
     if alg.initguess == :svd
-        z = _svd_guess(x, y, alg.D)
+        z = _svd_guess(x, y, alg.trunc.D)
     elseif alg.initguess == :rand
-        z = randomadt(promote_type(scalartype(x), scalartype(y)), phydims(x), D=alg.D)
+        z = randomadt(promote_type(scalartype(x), scalartype(y)), phydims(x), D=alg.trunc.D)
         canonicalize!(z, alg=Orthogonalize(normalize=true))
     elseif alg.initguess == :pre
-        z = increase_bond!(copy(x), alg.D)
+        z = increase_bond!(copy(x), alg.trunc.D)
         setscaling!(z, 1)
     else
         error("unsupported initguess $(alg.initguess)")