+++ src/defaults.jl
@@ -12,11 +12,8 @@ end
 
 const DefaultTruncation = truncdimcutoff(D=Defaults.D, ϵ=Defaults.tolgauge, add_back=0)
 
-
-const DefaultIntegrationTruncation = truncdimcutoff(D=10000, ϵ=1.0e-12, add_back=0)
 const DefaultITruncation = truncdimcutoff(D=200, ϵ=1.0e-10, add_back=0)
-const DefaultKTruncation = truncdimcutoff(D=1000, ϵ=1.0e-10, add_back=0)
-const DefaultMPOTruncation = truncdimcutoff(D=10000, ϵ=1.0e-14, add_back=0)
+const DefaultKTruncation = trunccutoff(Defaults.tolgauge) # system dynamics, initial-state absorption and MPO compression
 
 const DefaultMultAlg = DMRG1(DefaultITruncation) # default compression algorithm for MPS/MPO multiplication
 # const DefaultMultAlg = SVDCompression(DefaultITruncation)